+++ plugins/chart/demo/editor.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8" />
     <title>Editor</title>
     <link rel="stylesheet" href="https://uicdn.toast.com/chart/v4.1.4/toastui-chart.min.css" />
-    </style>
   </head>
   <body>
     <div class="code-html">
