+++ plugins/chart/package.json
@@ -36,12 +36,8 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.9",
-    "@toast-ui/editor": "^2.0.0",
-    "@types/highlight.js": "^9.12.4",
     "@types/jest": "^26.0.20",
-    "@types/lowlight": "0.0.1",
     "@types/node": "^14.14.31",
-    "@types/prosemirror-state": "^1.2.5",
     "@types/snowpack-env": "^2.3.3",
     "@typescript-eslint/eslint-plugin": "^4.16.1",
     "@typescript-eslint/parser": "^4.16.1",
